newbackend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
requests>=2.28.0
yt-dlp==2024.11.4
opencv-python==4.10.0.84
numpy==2.1.3
SQLAlchemy==2.0.36
pydantic-settings==2.6.1
faster-whisper==1.0.3
openai>=1.0.0,<2.0.0
easyocr>=1.6.0
Pillow>=9.0.0
psutil>=5.8.0
aiofiles>=23.0.0
imagehash>=4.3.0
